# lb6/Service/OpenAPI_service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/user/login": {
      "get": {
        "tags": [
          "Страница пользователя"
        ],
        "summary": "User Hello",
        "operationId": "user_hello_api_v1_user_login_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/project/create": {
      "post": {
        "tags": [
          "Проекты"
        ],
        "summary": "Create Project Route",
        "description": "Создает новый проект",
        "operationId": "create_project_route_api_v1_project_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/project/search": {
      "get": {
        "tags": [
          "Проекты"
        ],
        "summary": "Search Project By Name Route",
        "description": "Ищет проекты по имени",
        "operationId": "search_project_by_name_route_api_v1_project_search_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Проекты"
        ],
        "summary": "Get All Projects Route",
        "description": "Возвращает список всех проектов",
        "operationId": "get_all_projects_route_api_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/task/create": {
      "post": {
        "tags": [
          "Задачи"
        ],
        "summary": "Create Task Route",
        "description": "Создает новую задачу в проекте",
        "operationId": "create_task_route_api_v1_task_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Задачи"
        ],
        "summary": "Get All Tasks In Project Route",
        "description": "Возвращает список всех задач в проекте",
        "operationId": "get_all_tasks_in_project_route_api_v1_tasks_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/task": {
      "get": {
        "tags": [
          "Задачи"
        ],
        "summary": "Get Task By Code Route",
        "description": "Возвращает задачу по коду",
        "operationId": "get_task_by_code_route_api_v1_task_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Новый проект"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "example": "Описание нового проекта"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "title": "ProjectCreate"
      },
      "TaskCreate": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "example": 1
          },
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Новая задача"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "example": "Описание новой задачи"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "example": "open"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title",
          "description",
          "status"
        ],
        "title": "TaskCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}